implement base architecture and comments module/add table Cooments

+++ Controllers/AuthorControllers.cs
@@ -1,42 +1,101 @@
 ﻿using Homework2.DataAccess;
 using Homework2.Entities;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+using Homework2.DTOs;
+using AutoMapper;
+using Microsoft.AspNetCore.JsonPatch;
 
 namespace Homework2.Controllers
 {
-    [ApiController]        //Normally I use [Controller], but in this code I prefer to use [ApiController]
-                           // because with it I don't need to specify where the Post came from
+    //[ApiController]        //Normally I use [Controller], but in this code I prefer to use [ApiController]
+    // because with it I don't need to specify where the Post came from
     [Route("Api/Authors")]
-    public class AuthorControllers:ControllerBase
+    public class AuthorControllers : BaseController<Author, AuthorDTO>
     {
-        public readonly DataDbContext _DbContext;
-        public AuthorControllers(DataDbContext _DbContext)
+        public AuthorControllers(ApplicationDbContext context, ILogger<AuthorControllers> logger, IMapper mapper)
+            : base(mapper, logger, context)
+        {
+        }
+    
+        [HttpPatch("{Id:int}")]
+        public async Task<ActionResult> Patch(int id, JsonPatchDocument<AuthorPatchDTO> PatchDoc)
+        {
+            if (PatchDoc is null) return BadRequest();//400
+            
+
+            var author = _context.Authors.FirstOrDefault(a => a.Id == id);
+
+            var authorPatchDTO = _mapper.Map<AuthorPatchDTO>(author);//remain here i don't used hierarchy like Author to AuthorDTO to AuthorPatchDTO
+
+            if (author is null) return NotFound();//404
+            
+
+            PatchDoc.ApplyTo(authorPatchDTO, ModelState);
+
+            if (!TryValidateModel(authorPatchDTO)) return ValidationProblem();//400
+            
+            _mapper.Map(authorPatchDTO,author);
+
+            await _context.SaveChangesAsync();
+
+            return NoContent();//204
+        }
+    }
+}
+
+
+        /*
+        public readonly ApplicationDbContext _DbContext;
+        public readonly ILogger<AuthorControllers> logger;   // I used logger for practice, now not is needs
+        public readonly IMapper mapper;
+        public AuthorControllers(ApplicationDbContext _DbContext,ILogger<AuthorControllers> _logger, IMapper _mapper)
         {
             this._DbContext = _DbContext;
+            this.logger = _logger;
+            this.mapper = _mapper;
         }
 
 
 
         [HttpGet]
-        public async Task<List<Author>> Gets()
+        public async Task<IEnumerable<AuthorDTO>> GetAthors()
         {
-            return await _DbContext.Authors.ToListAsync<Author>();
-        }
+            var authors = await _DbContext.Authors.ToListAsync<Author>();
+            var authorsDTO = mapper.Map<IEnumerable<AuthorDTO>>(authors);
 
-
-
-        [HttpGet("{Id:int}")] //A Get for specify author with Api/Authors/#x
-        public async Task<ActionResult<Author>> Get(int id)
-        {
-            var author = await _DbContext.Authors.FirstOrDefaultAsync(a => a.Id == id);
-
-            if (author is null)
+            return authorsDTO;
+            
+            //---------------------------------------------------
+            //The way manual is like that but i used Pakege because is more fast and easy. Name of Library is AutoMapper
+            /*
+            var authorsDTO = authors.Select(authors => new AuthorDTO //authorsDTO is IEnumerable because Selcet and other function of Lambda always return IEnumerable
             {
-                return BadRequest("Author doesn't exist");
+                Id = authors.Id,
+                FullName = $"{authors.Name} {authors.LastName}"
+            });
+            return authorsDTO;
+            */
+        /*
+        }
+        */
+
+        /*
+
+        [HttpGet("{Id:int}", Name = "GetAuthor")] //A Get for specify author with Api/Authors/#x
+        public async Task<ActionResult<AuthorDTO>> GetAthor(int id)
+        {
+
+           
+            var author = await _DbContext.Authors.FirstOrDefaultAsync(a => a.Id == id);
+            var authorDTO = mapper.Map<AuthorDTO>(author);
+
+            if (authorDTO is null)
+            {
+                logger.LogWarning("The Author not found");
+                return NotFound("Author not found"); //404
             }
 
-            return author;
+            return authorDTO;
         }
 
         [HttpPost]
@@ -44,7 +103,7 @@ namespace Homework2.Controllers
         {
             _DbContext.Add(author);
             await _DbContext.SaveChangesAsync();
-            return Ok("Author Creado");
+            return CreatedAtRoute("GetAuthor", new {Id = author.Id}, author);//201
         }
 
 
@@ -53,25 +112,35 @@ namespace Homework2.Controllers
         {
             if(id != author.Id)
             {
-                return BadRequest("The Ids not match");
+                logger.LogWarning("Author Id and Id not macth ");
+                return BadRequest("The Ids not match");//401
             }
 
             _DbContext.Update(author);
             await _DbContext.SaveChangesAsync();
-            return Ok("Update Done Corretly");
+            return NoContent(); //204
         }
 
+      
+        //Note, no is normally use delete because is loser info, the commun is create atributte bolean like astive=true/false
+        //
         [HttpDelete("{Id:int}")]   //Delete specify Author with Api/Authors/#x
         public async Task<ActionResult> Delete(int id)
         {
+            var author = await _DbContext.Authors.FirstOrDefaultAsync(x => x.Id == id);
+            
+            if(author is null)
+            {
+                logger.LogWarning("Author not found whet user try to delete");
+                return NotFound("Not Found some Author With Id"); //401
+            }
             var result = await _DbContext.Authors.Where(a => a.Id == id).ExecuteDeleteAsync();
 
-            if(result == 0)
-            {
-                return NotFound("Not Found some Author With Id");
-            }
 
-            return Ok("Author Deleted Correctly");
+            logger.LogInformation("Author Deleted");
+            return NoContent();//204
         }
+        
     }
 }
+*/
