+++ Controllers/PublisherControllers.cs
@@ -1,20 +1,32 @@
-﻿using Homework2.DataAccess;
+﻿using AutoMapper;
+using Homework2.DataAccess;
+using Homework2.DTOs;
 using Homework2.Entities;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
 namespace Homework2.Controllers
 {
-    [ApiController]    //Normally I use [Controller], but in this code I prefer to use [ApiController]
+    //[ApiController]    //Normally I use [Controller], but in this code I prefer to use [ApiController]
                        // because with it I don't need to specify where the Post came from
 
     [Route("Api/Publishers")]
-    public class PublisherControllers: ControllerBase
+    public class PublisherControllers: BaseController<Publisher,PublisherDTO>
     {
-        public readonly DataDbContext _DbContext;
-        public PublisherControllers(DataDbContext _DbContext)
+
+        public PublisherControllers(ApplicationDbContext context, ILogger<PublisherControllers> logger, IMapper mapper)
+            :base(mapper,logger,context)
+        {
+        }
+
+        //now i watch that this Controller don't had Mapper
+        /*
+        public readonly ApplicationDbContext _DbContext;
+        public readonly ILogger<LoanControllers> logger;
+        public PublisherControllers(ApplicationDbContext _DbContext, ILogger<LoanControllers> _logger)
         {
             this._DbContext = _DbContext;
+            this.logger = _logger;
         }
 
         [HttpGet]
@@ -24,14 +36,14 @@ namespace Homework2.Controllers
         }
 
 
-        [HttpGet("{Id:int}")]  //A Get for specify Publisher with Api/Publishers/#x
+        [HttpGet("{Id:int}", Name = "GetPublisher")]  //A Get for specify Publisher with Api/Publishers/#x
         public async Task<ActionResult<Publisher>> Get(int id)
         {
             var publisher = await _DbContext.Publishers.FirstOrDefaultAsync(a => a.Id == id);
 
             if (publisher is null)
             {
-                return BadRequest("Author doesn't exist");
+                return BadRequest("Publihsher doesn't exist");//401
             }
 
             return publisher;
@@ -42,7 +54,8 @@ namespace Homework2.Controllers
         {
             _DbContext.Add(publisher);
             await _DbContext.SaveChangesAsync();
-            return Ok("Publisher Creado");
+            logger.LogInformation("Publisher Created");
+            return CreatedAtRoute("GetPublisher", new {id = publisher.Id}, publisher);//201
         }
 
 
@@ -51,25 +64,28 @@ namespace Homework2.Controllers
         {
             if (id != publisher.Id)
             {
-                return BadRequest("The Ids not match");
+                return BadRequest("The Ids not match");//401
             }
 
             _DbContext.Update(publisher);
             await _DbContext.SaveChangesAsync();
-            return Ok("Update Done Corretly");
+            logger.LogInformation("Publisher Update");
+            return CreatedAtRoute("GetPublisher", new {id = publisher.Id}, publisher);//201
         }
 
         [HttpDelete("{Id:int}")] //A DElete for specify Publisher with Api/Publishers/#x
         public async Task<ActionResult> Delete(int id)
         {
-            var result = await _DbContext.Publishers.Where(a => a.Id == id).ExecuteDeleteAsync();
+            var publisher = _DbContext.Publishers.FirstOrDefaultAsync(x => x.Id == id);
 
-            if (result == 0)
+            if (publisher is null)
             {
-                return NotFound("Not Found some Publisher With Id");
+                return NotFound("Not Found some Publisher With Id");//404
             }
-
-            return Ok("Publisher Deleted Correctly");
+            var result = await _DbContext.Publishers.Where(a => a.Id == id).ExecuteDeleteAsync();
+            logger.LogInformation("Publisher deleted");
+            return CreatedAtRoute("GetPublisher", new {id = publisher.Id}, publisher);//201
         }
+        */
     }
 }
